Pruebas basicas y estructuracion de react
